optimize the shielded TRC20 code
