+++ framework/src/main/java/org/tron/core/services/http/IsShieldedTRC20ContractNoteSpentServlet.java
@@ -19,7 +19,6 @@ public class IsShieldedTRC20ContractNoteSpentServlet extends RateLimiterServlet 
   private Wallet wallet;
 
   protected void doGet(HttpServletRequest request, HttpServletResponse response) {
-
   }
 
   protected void doPost(HttpServletRequest request, HttpServletResponse response) {
@@ -28,10 +27,8 @@ public class IsShieldedTRC20ContractNoteSpentServlet extends RateLimiterServlet 
           .collect(Collectors.joining(System.lineSeparator()));
       Util.checkBodySize(input);
       boolean visible = Util.getVisiblePost(input);
-
       NfTRC20Parameters.Builder build = NfTRC20Parameters.newBuilder();
       JsonFormat.merge(input, build);
-
       GrpcAPI.NullifierResult result = wallet.isShieldedTRC20ContractNoteSpent(build.build());
       response.getWriter().println(JsonFormat.printToString(result, visible));
     } catch (Exception e) {
