+++ framework/src/main/java/org/tron/core/zen/ShieldedTRC20ParametersBuilder.java
@@ -12,28 +12,24 @@ import org.apache.commons.lang3.ArrayUtils;
 import org.tron.api.GrpcAPI.ShieldedTRC20Parameters;
 import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.ByteUtil;
-import org.tron.common.utils.DBConfig;
 import org.tron.common.utils.Sha256Hash;
 import org.tron.common.zksnark.JLibrustzcash;
 import org.tron.common.zksnark.LibrustzcashParam;
 import org.tron.common.zksnark.ZksnarkUtils;
-import org.tron.core.Wallet;
 import org.tron.core.capsule.ReceiveDescriptionCapsule;
 import org.tron.core.capsule.SpendDescriptionCapsule;
-import org.tron.core.capsule.TransactionCapsule;
 import org.tron.core.exception.ZksnarkException;
-import org.tron.core.utils.TransactionUtil;
 import org.tron.core.zen.address.DiversifierT;
 import org.tron.core.zen.address.ExpandedSpendingKey;
 import org.tron.core.zen.address.PaymentAddress;
 import org.tron.core.zen.note.Note;
 import org.tron.core.zen.note.NoteEncryption;
 import org.tron.core.zen.note.OutgoingPlaintext;
-import org.tron.protos.Protocol;
 import org.tron.protos.contract.ShieldContract;
 
 @Slf4j
 public class ShieldedTRC20ParametersBuilder {
+
   private static final int MERKLE_TREE_PATH_LENGTH = 1024; // 32*32
   private static final String MERKLE_TREE_PATH_LENGTH_ERROR = "Merkle tree path format is wrong";
   @Setter
@@ -65,17 +61,17 @@ public class ShieldedTRC20ParametersBuilder {
     result[0] = 0x20;
     for (int i = 0; i < 32; i++) {
       result[1 + i * 33] = 0x20;
-      System.arraycopy(path, i * 32, result, 2 + i * 33,32);
+      System.arraycopy(path, i * 32, result, 2 + i * 33, 32);
     }
     byte[] positionBytes = ByteArray.fromLong(position);
     ZksnarkUtils.sort(positionBytes);
-    System.arraycopy(positionBytes, 0, result, 1057,8);
+    System.arraycopy(positionBytes, 0, result, 1057, 8);
     return result;
   }
 
   // Note: should call librustzcashSaplingProvingCtxFree in the caller
   private SpendDescriptionCapsule generateSpendProof(SpendDescriptionInfo spend,
-            long ctx) throws ZksnarkException {
+      long ctx) throws ZksnarkException {
     byte[] cm = spend.note.cm();
     //check if ak exists
     byte[] ak;
@@ -98,18 +94,18 @@ public class ShieldedTRC20ParametersBuilder {
     byte[] rk = new byte[32];
     byte[] zkproof = new byte[192];
     if (!JLibrustzcash.librustzcashSaplingSpendProof(
-                new LibrustzcashParam.SpendProofParams(ctx,
-                        ak,
-                        nsk,
-                        spend.note.getD().getData(),
-                        spend.note.getRcm(),
-                        spend.alpha,
-                        spend.note.getValue(),
-                        spend.anchor,
-                        path,
-                        cv,
-                        rk,
-                        zkproof))) {
+        new LibrustzcashParam.SpendProofParams(ctx,
+            ak,
+            nsk,
+            spend.note.getD().getData(),
+            spend.note.getRcm(),
+            spend.alpha,
+            spend.note.getValue(),
+            spend.anchor,
+            path,
+            cv,
+            rk,
+            zkproof))) {
       throw new ZksnarkException("Spend proof failed");
     }
     SpendDescriptionCapsule spendDescriptionCapsule = new SpendDescriptionCapsule();
@@ -123,13 +119,13 @@ public class ShieldedTRC20ParametersBuilder {
 
   // Note: should call librustzcashSaplingProvingCtxFree in the caller
   private ReceiveDescriptionCapsule generateOutputProof(ReceiveDescriptionInfo output, long ctx)
-            throws ZksnarkException {
+      throws ZksnarkException {
     byte[] cm = output.getNote().cm();
     if (ByteArray.isEmpty(cm)) {
       throw new ZksnarkException("Output is invalid");
     }
     Optional<Note.NotePlaintextEncryptionResult> res = output.getNote()
-                .encrypt(output.getNote().getPkD());
+        .encrypt(output.getNote().getPkD());
     if (!res.isPresent()) {
       throw new ZksnarkException("Failed to encrypt note");
     }
@@ -138,14 +134,14 @@ public class ShieldedTRC20ParametersBuilder {
     byte[] cv = new byte[32];
     byte[] zkProof = new byte[192];
     if (!JLibrustzcash.librustzcashSaplingOutputProof(
-                new LibrustzcashParam.OutputProofParams(ctx,
-                        encryptor.getEsk(),
-                        output.getNote().getD().getData(),
-                        output.getNote().getPkD(),
-                        output.getNote().getRcm(),
-                        output.getNote().getValue(),
-                        cv,
-                        zkProof))) {
+        new LibrustzcashParam.OutputProofParams(ctx,
+            encryptor.getEsk(),
+            output.getNote().getD().getData(),
+            output.getNote().getPkD(),
+            output.getNote().getRcm(),
+            output.getNote().getValue(),
+            cv,
+            zkProof))) {
       throw new ZksnarkException("Output proof failed");
     }
     if (ArrayUtils.isEmpty(output.ovk) || output.ovk.length != 32) {
@@ -158,11 +154,11 @@ public class ShieldedTRC20ParametersBuilder {
     receiveDescriptionCapsule.setCEnc(enc.getEncCiphertext());
     receiveDescriptionCapsule.setZkproof(zkProof);
     OutgoingPlaintext outPlaintext =
-                new OutgoingPlaintext(output.getNote().getPkD(), encryptor.getEsk());
+        new OutgoingPlaintext(output.getNote().getPkD(), encryptor.getEsk());
     receiveDescriptionCapsule.setCOut(outPlaintext
-                .encrypt(output.ovk, receiveDescriptionCapsule.getValueCommitment().toByteArray(),
-                        receiveDescriptionCapsule.getCm().toByteArray(),
-                        encryptor).getData());
+        .encrypt(output.ovk, receiveDescriptionCapsule.getValueCommitment().toByteArray(),
+            receiveDescriptionCapsule.getCm().toByteArray(),
+            encryptor).getData());
     return receiveDescriptionCapsule;
   }
 
@@ -170,37 +166,37 @@ public class ShieldedTRC20ParametersBuilder {
     for (int i = 0; i < spends.size(); i++) {
       byte[] result = new byte[64];
       JLibrustzcash.librustzcashSaplingSpendSig(
-                    new LibrustzcashParam.SpendSigParams(spends.get(i).expsk.getAsk(),
-                            spends.get(i).alpha,
-                            dataToBeSigned,
-                            result));
+          new LibrustzcashParam.SpendSigParams(spends.get(i).expsk.getAsk(),
+              spends.get(i).alpha,
+              dataToBeSigned,
+              result));
       builder.getSpendDescriptionBuilder(i)
-                    .setSpendAuthoritySignature(ByteString.copyFrom(result));
+          .setSpendAuthoritySignature(ByteString.copyFrom(result));
     }
   }
 
   private byte[] encodeSpendDescriptionWithoutSpendAuthSig(
-          ShieldContract.SpendDescription spendDescription) {
+      ShieldContract.SpendDescription spendDescription) {
     return ByteUtil.merge(spendDescription.getNullifier().toByteArray(),
-                              spendDescription.getAnchor().toByteArray(),
-                              spendDescription.getValueCommitment().toByteArray(),
-                              spendDescription.getRk().toByteArray(),
-                              spendDescription.getZkproof().toByteArray());
+        spendDescription.getAnchor().toByteArray(),
+        spendDescription.getValueCommitment().toByteArray(),
+        spendDescription.getRk().toByteArray(),
+        spendDescription.getZkproof().toByteArray());
   }
 
   private byte[] encodeReceiveDescriptionWithoutC(
-          ShieldContract.ReceiveDescription receiveDescription) {
+      ShieldContract.ReceiveDescription receiveDescription) {
     return ByteUtil.merge(receiveDescription.getNoteCommitment().toByteArray(),
-                              receiveDescription.getValueCommitment().toByteArray(),
-                              receiveDescription.getEpk().toByteArray(),
-                              receiveDescription.getZkproof().toByteArray());
+        receiveDescription.getValueCommitment().toByteArray(),
+        receiveDescription.getEpk().toByteArray(),
+        receiveDescription.getZkproof().toByteArray());
   }
 
   private byte[] encodeCencCout(ShieldContract.ReceiveDescription receiveDescription) {
     byte[] padding = new byte[12];
     return ByteUtil.merge(receiveDescription.getCEnc().toByteArray(),
-                receiveDescription.getCOut().toByteArray(),
-                padding);
+        receiveDescription.getCOut().toByteArray(),
+        padding);
   }
 
   public ShieldedTRC20Parameters build(boolean withAsk) throws ZksnarkException {
@@ -213,14 +209,13 @@ public class ShieldedTRC20ParametersBuilder {
 
     switch (shieldedTRC20ParametersType) {
       case MINT:
-        ReceiveDescriptionInfo receive =  receives.get(0);
+        ReceiveDescriptionInfo receive = receives.get(0);
         receiveDescription = generateOutputProof(receive, ctx).getInstance();
         builder.addReceiveDescription(receiveDescription);
-        //valueBalance += transparentFromAmount;
         mergedBytes = ByteUtil.merge(shieldedTRC20Address, ByteArray
-                        .fromLong(transparentFromAmount),
-                        encodeReceiveDescriptionWithoutC(receiveDescription),
-                encodeCencCout(receiveDescription));
+                .fromLong(transparentFromAmount),
+            encodeReceiveDescriptionWithoutC(receiveDescription),
+            encodeCencCout(receiveDescription));
         break;
       case TRANSFER:
         // Create SpendDescriptions
@@ -229,7 +224,7 @@ public class ShieldedTRC20ParametersBuilder {
           spendDescription = generateSpendProof(spend, ctx).getInstance();
           builder.addSpendDescription(spendDescription);
           mergedBytes = ByteUtil.merge(mergedBytes,
-                  encodeSpendDescriptionWithoutSpendAuthSig(spendDescription));
+              encodeSpendDescriptionWithoutSpendAuthSig(spendDescription));
         }
         // Create OutputDescriptions
         byte[] cencCout = new byte[0];
@@ -237,7 +232,7 @@ public class ShieldedTRC20ParametersBuilder {
           receiveDescription = generateOutputProof(receiveD, ctx).getInstance();
           builder.addReceiveDescription(receiveDescription);
           mergedBytes = ByteUtil.merge(mergedBytes,
-                  encodeReceiveDescriptionWithoutC(receiveDescription));
+              encodeReceiveDescriptionWithoutC(receiveDescription));
           cencCout = ByteUtil.merge(cencCout, encodeCencCout(receiveDescription));
         }
         mergedBytes = ByteUtil.merge(mergedBytes, cencCout);
@@ -246,19 +241,17 @@ public class ShieldedTRC20ParametersBuilder {
         SpendDescriptionInfo spend = spends.get(0);
         spendDescription = generateSpendProof(spend, ctx).getInstance();
         builder.addSpendDescription(spendDescription);
-        //valueBalance -= transparentToAmount;
         mergedBytes = ByteUtil.merge(shieldedTRC20Address,
-                encodeSpendDescriptionWithoutSpendAuthSig(spendDescription),
-                                    transparentToAddress,ByteArray.fromLong(transparentToAmount));
+            encodeSpendDescriptionWithoutSpendAuthSig(spendDescription),
+            transparentToAddress, ByteArray.fromLong(transparentToAmount));
         break;
       default:
         mergedBytes = null;
     }
-    dataHashToBeSigned = Sha256Hash.of(true,mergedBytes).getBytes();
+    dataHashToBeSigned = Sha256Hash.of(true, mergedBytes).getBytes();
     if (dataHashToBeSigned == null) {
       throw new ZksnarkException("cal transaction hash failed");
     }
-    // Create spendAuth and binding signatures
     if (withAsk) {
       createSpendAuth(dataHashToBeSigned);
     }
@@ -266,10 +259,10 @@ public class ShieldedTRC20ParametersBuilder {
     try {
       byte[] bindingSig = new byte[64];
       JLibrustzcash.librustzcashSaplingBindingSig(
-                    new LibrustzcashParam.BindingSigParams(ctx,
-                            valueBalance,
-                            dataHashToBeSigned,
-                            bindingSig)
+          new LibrustzcashParam.BindingSigParams(ctx,
+              valueBalance,
+              dataHashToBeSigned,
+              bindingSig)
       );
       builder.setBindingSignature(ByteString.copyFrom(bindingSig));
     } catch (ZksnarkException e) {
@@ -281,41 +274,41 @@ public class ShieldedTRC20ParametersBuilder {
   }
 
   public void addSpend(
-            ExpandedSpendingKey expsk,
-            Note note,
-            byte[] anchor,
-            byte[] path,
-            long position) throws ZksnarkException {
+      ExpandedSpendingKey expsk,
+      Note note,
+      byte[] anchor,
+      byte[] path,
+      long position) throws ZksnarkException {
     spends.add(new SpendDescriptionInfo(expsk, note, anchor, path, position));
     valueBalance += note.getValue();
   }
 
   public void addSpend(
-            ExpandedSpendingKey expsk,
-            Note note,
-            byte[] alpha,
-            byte[] anchor,
-            byte[] path,
-            long position) {
+      ExpandedSpendingKey expsk,
+      Note note,
+      byte[] alpha,
+      byte[] anchor,
+      byte[] path,
+      long position) {
     spends.add(new SpendDescriptionInfo(expsk, note, alpha, anchor, path, position));
     valueBalance += note.getValue();
   }
 
   public void addSpend(
-            byte[] ak,
-            byte[] nsk,
-            byte[] ovk,
-            Note note,
-            byte[] alpha,
-            byte[] anchor,
-            byte[] path,
-            long position) {
+      byte[] ak,
+      byte[] nsk,
+      byte[] ovk,
+      Note note,
+      byte[] alpha,
+      byte[] anchor,
+      byte[] path,
+      long position) {
     spends.add(new SpendDescriptionInfo(ak, nsk, ovk, note, alpha, anchor, path, position));
     valueBalance += note.getValue();
   }
 
   public void addOutput(byte[] ovk, PaymentAddress to, long value, byte[] memo)
-            throws ZksnarkException {
+      throws ZksnarkException {
     Note note = new Note(to, value);
     note.setMemo(memo);
     receives.add(new ReceiveDescriptionInfo(ovk, note));
@@ -330,6 +323,7 @@ public class ShieldedTRC20ParametersBuilder {
   }
 
   public static class SpendDescriptionInfo {
+
     @Getter
     @Setter
     private ExpandedSpendingKey expsk;
@@ -359,11 +353,11 @@ public class ShieldedTRC20ParametersBuilder {
     private byte[] ovk;
 
     private SpendDescriptionInfo(
-                ExpandedSpendingKey expsk,
-                Note note,
-                byte[] anchor,
-                byte[] path,
-                long position) throws ZksnarkException {
+        ExpandedSpendingKey expsk,
+        Note note,
+        byte[] anchor,
+        byte[] path,
+        long position) throws ZksnarkException {
       this.expsk = expsk;
       this.note = note;
       this.anchor = anchor;
@@ -374,12 +368,12 @@ public class ShieldedTRC20ParametersBuilder {
     }
 
     private SpendDescriptionInfo(
-                ExpandedSpendingKey expsk,
-                Note note,
-                byte[] alpha,
-                byte[] anchor,
-                byte[] path,
-                long position) {
+        ExpandedSpendingKey expsk,
+        Note note,
+        byte[] alpha,
+        byte[] anchor,
+        byte[] path,
+        long position) {
       this.expsk = expsk;
       this.note = note;
       this.anchor = anchor;
@@ -389,14 +383,14 @@ public class ShieldedTRC20ParametersBuilder {
     }
 
     private SpendDescriptionInfo(
-                byte[] ak,
-                byte[] nsk,
-                byte[] ovk,
-                Note note,
-                byte[] alpha,
-                byte[] anchor,
-                byte[] path,
-                long position) {
+        byte[] ak,
+        byte[] nsk,
+        byte[] ovk,
+        Note note,
+        byte[] alpha,
+        byte[] anchor,
+        byte[] path,
+        long position) {
       this.ak = ak;
       this.nsk = nsk;
       this.note = note;
@@ -410,6 +404,7 @@ public class ShieldedTRC20ParametersBuilder {
 
   @AllArgsConstructor
   private class ReceiveDescriptionInfo {
+
     @Getter
     private byte[] ovk;
     @Getter
@@ -417,8 +412,8 @@ public class ShieldedTRC20ParametersBuilder {
   }
 
   public enum ShieldedTRC20ParametersType {
-        MINT,
-        TRANSFER,
-        BURN
+    MINT,
+    TRANSFER,
+    BURN
   }
 }
