+++ framework/src/test/java/org/tron/core/WalletTest.java
@@ -21,42 +21,27 @@ package org.tron.core;
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static org.tron.core.zksnark.LibrustzcashTest.librustzcashInitZksnarkParams;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.ByteString;
 import java.io.File;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-
-import io.grpc.ManagedChannel;
-import io.grpc.ManagedChannelBuilder;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.ArrayUtils;
 import org.joda.time.DateTime;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
-import org.spongycastle.util.encoders.Hex;
-import org.tron.api.GrpcAPI;
 import org.tron.api.GrpcAPI.AssetIssueList;
 import org.tron.api.GrpcAPI.BlockList;
 import org.tron.api.GrpcAPI.ExchangeList;
 import org.tron.api.GrpcAPI.ProposalList;
-import org.tron.api.WalletGrpc;
 import org.tron.common.application.TronApplicationContext;
 import org.tron.common.crypto.ECKey;
-import org.tron.common.utils.*;
-import org.tron.common.zksnark.JLibrustzcash;
-import org.tron.common.zksnark.LibrustzcashParam;
+import org.tron.common.utils.ByteArray;
+import org.tron.common.utils.FileUtil;
+import org.tron.common.utils.Utils;
 import org.tron.core.capsule.AssetIssueCapsule;
 import org.tron.core.capsule.BlockCapsule;
 import org.tron.core.capsule.ExchangeCapsule;
@@ -66,14 +51,8 @@ import org.tron.core.capsule.TransactionInfoCapsule;
 import org.tron.core.config.DefaultConfig;
 import org.tron.core.config.args.Args;
 import org.tron.core.db.Manager;
-import org.tron.core.exception.ContractValidateException;
-import org.tron.core.exception.VMIllegalException;
-import org.tron.core.exception.ZksnarkException;
 import org.tron.core.store.DynamicPropertiesStore;
 import org.tron.core.utils.ProposalUtil.ProposalType;
-import org.tron.core.vm.PrecompiledContracts;
-import org.tron.core.zen.address.*;
-import org.tron.core.zen.note.Note;
 import org.tron.protos.Protocol;
 import org.tron.protos.Protocol.Block;
 import org.tron.protos.Protocol.BlockHeader;
@@ -86,10 +65,6 @@ import org.tron.protos.Protocol.Transaction.Contract.ContractType;
 import org.tron.protos.Protocol.TransactionInfo;
 import org.tron.protos.contract.AssetIssueContractOuterClass.AssetIssueContract;
 import org.tron.protos.contract.BalanceContract.TransferContract;
-import org.tron.protos.contract.ShieldContract;
-import org.tron.protos.contract.SmartContractOuterClass;
-import stest.tron.wallet.common.client.WalletClient;
-import stest.tron.wallet.common.client.utils.PublicMethed;
 
 //import org.tron.protos.Protocol.DeferredTransaction;
 
@@ -538,1857 +513,5 @@ public class WalletTest {
     });
 
     System.out.printf(builder.build().toString());
-  }
-
-  private GrpcAPI.Note getNote(long value, String paymentAddress, byte[] rcm, byte[] memo) {
-    GrpcAPI.Note.Builder noteBuilder = GrpcAPI.Note.newBuilder();
-    noteBuilder.setValue(value);
-    noteBuilder.setPaymentAddress(paymentAddress);
-    noteBuilder.setRcm(ByteString.copyFrom(rcm));
-    noteBuilder.setMemo(ByteString.copyFrom(memo));
-    return noteBuilder.build();
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForMint()
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-
-    long from_amount = 50;
-    SpendingKey sk = SpendingKey.random();
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    byte[] ovk = expsk.getOvk();
-
-    //ReceiveNote
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.decode(privateKey);
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 50;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    byte[] contractAddress = WalletClient.decodeFromBase58Check(getContractAddress());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder paramBuilder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    paramBuilder.setOvk(ByteString.copyFrom(ovk));
-    paramBuilder.setFromAmount(from_amount);
-    paramBuilder.addShieldedReceives(revNoteBuilder.build());
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-
-    GrpcAPI.ShieldedTRC20Parameters trc20MintParams = blockingStubFull
-            .createShieldedContractParameters(paramBuilder.build());
-    GrpcAPI.PrivateShieldedTRC20Parameters trc20Params = paramBuilder.build();
-    logger.info(Hex.toHexString(trc20Params.getOvk().toByteArray()));
-    logger.info(String.valueOf(trc20Params.getFromAmount()));
-    logger.info(String.valueOf(trc20Params.getShieldedReceives(0).getNote().getValue()));
-    logger.info(trc20Params.getShieldedReceives(0).getNote().getPaymentAddress());
-    logger.info(Hex.toHexString(
-            trc20Params.getShieldedReceives(0).getNote().getRcm().toByteArray()));
-    logger.info(Hex.toHexString(trc20Params.getShieldedTRC20ContractAddress().toByteArray()));
-
-    //verify receiveProof && bindingSignature
-    boolean result;
-    long ctx = JLibrustzcash.librustzcashSaplingVerificationCtxInit();
-    ShieldContract.ReceiveDescription revDesc = trc20MintParams.getReceiveDescription(0);
-    try {
-      result = JLibrustzcash.librustzcashSaplingCheckOutput(
-              new LibrustzcashParam.CheckOutputParams(
-                      ctx,
-                      revDesc.getValueCommitment().toByteArray(),
-                      revDesc.getNoteCommitment().toByteArray(),
-                      revDesc.getEpk().toByteArray(),
-                      revDesc.getZkproof().toByteArray()));
-      long valueBalance = -revValue;
-      result &= JLibrustzcash.librustzcashSaplingFinalCheck(
-              new LibrustzcashParam.FinalCheckParams(
-                      ctx,
-                      valueBalance,
-                      trc20MintParams.getBindingSignature().toByteArray(),
-                      trc20MintParams.getMessageHash().toByteArray()));
-    } catch (Throwable any) {
-      result = false;
-    } finally {
-      JLibrustzcash.librustzcashSaplingVerificationCtxFree(ctx);
-    }
-    Assert.assertTrue(result);
-    String mintInput = mintParamsToHexString(trc20MintParams, revValue);
-    String txid1 = triggerMint(blockingStubFull, contractAddress, callerAddress, privateKey, mintInput);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-  }
-
-  private GrpcAPI.PrivateShieldedTRC20Parameters mintParams(String privKey,
-                                                            long value, String contractAddr)
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    long from_amount = value;
-    SpendingKey sk = SpendingKey.decode(privKey);
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    byte[] ask = expsk.getAsk();
-    byte[] nsk = expsk.getNsk();
-    byte[] ovk = expsk.getOvk();
-
-    // ReceiveNote
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    // SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = sk.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = value;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-    byte[] contractAddress = WalletClient.decodeFromBase58Check(contractAddr);
-
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder paramBuilder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    paramBuilder.setAsk(ByteString.copyFrom(ask));
-    paramBuilder.setNsk(ByteString.copyFrom(nsk));
-    paramBuilder.setOvk(ByteString.copyFrom(ovk));
-    paramBuilder.setFromAmount(from_amount);
-    paramBuilder.addShieldedReceives(revNoteBuilder.build());
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    return paramBuilder.build();
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransfer2v2()
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-    // mintPrivateParam1);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam2 = mintParams(
-            privateKey, 40, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam2 = wallet.createShieldedContractParameters(
-    // mintPrivateParam2);
-    GrpcAPI.ShieldedTRC20Parameters mintParam2 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam2);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String mintInput2 = mintParamsToHexString(mintParam2, 40);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    String txid2 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput2);
-    logger.info("..............mint result...........");
-    logger.info(txid1);
-    logger.info(txid2);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    // SpendNoteTRC20 2
-    Optional<TransactionInfo> infoById2 = PublicMethed
-            .getTransactionInfoById(txid2, blockingStubFull);
-    byte[] tx2Data = infoById2.get().getLog(0).getData().toByteArray();
-    long pos2 = bytes32Tolong(ByteArray.subArray(tx2Data, 0, 32));
-    String txid4 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos2);
-    Optional<TransactionInfo> infoById4 = PublicMethed
-            .getTransactionInfoById(txid4, blockingStubFull);
-    byte[] contractResult2 = infoById4.get().getContractResult(0).toByteArray();
-    byte[] path2 = ByteArray.subArray(contractResult2, 32, 1056);
-    byte[] root2 = ByteArray.subArray(contractResult2, 0, 32);
-    logger.info(Hex.toHexString(contractResult2));
-    GrpcAPI.SpendNoteTRC20.Builder note2Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note2Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note2Builder.setPos(pos2);
-    note2Builder.setPath(ByteString.copyFrom(path2));
-    note2Builder.setRoot(ByteString.copyFrom(root2));
-    note2Builder.setNote(mintPrivateParam2.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    privateTRC20Builder.addShieldedSpends(note2Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 30;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    //ReceiveNote 2
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder2 = GrpcAPI.ReceiveNote.newBuilder();
-    FullViewingKey fullViewingKey2 = sk.fullViewingKey();
-    IncomingViewingKey incomingViewingKey2 = fullViewingKey2.inViewingKey();
-    PaymentAddress paymentAddress2 = incomingViewingKey2.address(DiversifierT.random()).get();
-    long revValue2 = 70;
-    byte[] memo2 = new byte[512];
-    byte[] rcm2 = Note.generateR();
-    String paymentAddressStr2 = KeyIo.encodePaymentAddress(paymentAddress2);
-    GrpcAPI.Note revNote2 = getNote(revValue2, paymentAddressStr2, rcm2, memo2);
-    revNoteBuilder2.setNote(revNote2);
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder2.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    //GrpcAPI.ShieldedTRC20Parameters transferParam = wallet.createShieldedContractParameters(
-    // privateTRC20Builder.build());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull.createShieldedContractParameters(
-            privateTRC20Builder.build());
-    String transferInput = transferParamsToHexString(transferParam);
-    String txid = triggerTransfer(
-            blockingStubFull, contractAddress, callerAddress, privateKey, transferInput);
-    logger.info("..............transfer result...........");
-    logger.info(txid);
-    logger.info("..............end..............");
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransfer1v1()
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-    // mintPrivateParam1);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............mint result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 60;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-
-    //GrpcAPI.ShieldedTRC20Parameters transferParam = wallet.createShieldedContractParameters(
-    // privateTRC20Builder.build());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParameters(privateTRC20Builder.build());
-    String transferInput = transferParamsToHexString(transferParam);
-    String txid = triggerTransfer(
-            blockingStubFull, contractAddress, callerAddress, privateKey, transferInput);
-    logger.info("..............transfer result...........");
-    logger.info(txid);
-    logger.info("..............end..............");
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransfer1v2()
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 100, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-    // mintPrivateParam1);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 100);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............mint result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 30;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    //ReceiveNote 2
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder2 = GrpcAPI.ReceiveNote.newBuilder();
-    FullViewingKey fullViewingKey2 = sk.fullViewingKey();
-    IncomingViewingKey incomingViewingKey2 = fullViewingKey2.inViewingKey();
-    PaymentAddress paymentAddress2 = incomingViewingKey2.address(DiversifierT.random()).get();
-    long revValue2 = 70;
-    byte[] memo2 = new byte[512];
-    byte[] rcm2 = Note.generateR();
-    String paymentAddressStr2 = KeyIo.encodePaymentAddress(paymentAddress2);
-    GrpcAPI.Note revNote2 = getNote(revValue2, paymentAddressStr2, rcm2, memo2);
-    revNoteBuilder2.setNote(revNote2);
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder2.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-
-    //GrpcAPI.ShieldedTRC20Parameters transferParam = wallet.createShieldedContractParameters(
-    // privateTRC20Builder.build());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParameters(privateTRC20Builder.build());
-    String transferInput = transferParamsToHexString(transferParam);
-    String txid = triggerTransfer(
-            blockingStubFull, contractAddress, callerAddress, privateKey, transferInput);
-    logger.info("..............transfer result...........");
-    logger.info(txid);
-    logger.info("..............end..............");
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransfer2v1()
-          throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-    // mintPrivateParam1);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam2 = mintParams(
-            privateKey, 40, contractAddr);
-    //GrpcAPI.ShieldedTRC20Parameters mintParam2 = wallet.createShieldedContractParameters(
-    // mintPrivateParam2);
-    GrpcAPI.ShieldedTRC20Parameters mintParam2 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam2);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String mintInput2 = mintParamsToHexString(mintParam2, 40);
-    String txid1 = triggerMint(blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    String txid2 = triggerMint(blockingStubFull, contractAddress, callerAddress, privateKey, mintInput2);
-    logger.info("..............mint result...........");
-    logger.info(txid1);
-    logger.info(txid2);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    // SpendNoteTRC20 2
-    Optional<TransactionInfo> infoById2 = PublicMethed
-            .getTransactionInfoById(txid2, blockingStubFull);
-    byte[] tx2Data = infoById2.get().getLog(0).getData().toByteArray();
-    long pos2 = bytes32Tolong(ByteArray.subArray(tx2Data, 0, 32));
-    String txid4 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos2);
-    Optional<TransactionInfo> infoById4 = PublicMethed
-            .getTransactionInfoById(txid4, blockingStubFull);
-    byte[] contractResult2 = infoById4.get().getContractResult(0).toByteArray();
-    byte[] path2 = ByteArray.subArray(contractResult2, 32, 1056);
-    byte[] root2 = ByteArray.subArray(contractResult2, 0, 32);
-    logger.info(Hex.toHexString(contractResult2));
-    GrpcAPI.SpendNoteTRC20.Builder note2Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note2Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note2Builder.setPos(pos2);
-    note2Builder.setPath(ByteString.copyFrom(path2));
-    note2Builder.setRoot(ByteString.copyFrom(root2));
-    note2Builder.setNote(mintPrivateParam2.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    privateTRC20Builder.addShieldedSpends(note2Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 100;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-
-    //GrpcAPI.ShieldedTRC20Parameters transferParam = wallet.createShieldedContractParameters(
-    // privateTRC20Builder.build());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParameters(privateTRC20Builder.build());
-    String transferInput = transferParamsToHexString(transferParam);
-    String txid = triggerTransfer(
-            blockingStubFull, contractAddress, callerAddress, privateKey, transferInput);
-    logger.info("..............transfer result...........");
-    logger.info(txid);
-    logger.info("..............end..............");
-  }
-
-
-  private String triggerGetPath(WalletGrpc.WalletBlockingStub blockingStubFull, byte[] contractAddress,
-                                byte[] callerAddress, String privateKey, long pos) {
-    String txid = PublicMethed.triggerContract(contractAddress,
-            "getPath(uint256)",
-            Hex.toHexString(longTo32Bytes(pos)),
-            true,
-            0L,
-            1000000000L,
-            callerAddress, privateKey, blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Optional<TransactionInfo> infoById = PublicMethed
-            .getTransactionInfoById(txid, blockingStubFull);
-    logger.info("Trigger energytotal is " + infoById.get().getReceipt().getEnergyUsageTotal());
-    Assert.assertEquals(
-            Transaction.Result.contractResult.SUCCESS, infoById.get().getReceipt().getResult());
-    return txid;
-  }
-
-  private String triggerBurn(WalletGrpc.WalletBlockingStub blockingStubFull, byte[] contractAddress,
-                             byte[] callerAddress, String privateKey, String input) {
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    String txid = PublicMethed.triggerContract(contractAddress,
-            "burn(bytes32[10],bytes32[2],uint64,bytes32[2],uint256)",
-            input,
-            true,
-            0L, 1000000000L,
-            callerAddress, privateKey,
-            blockingStubFull);
-    // logger.info(txid);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Optional<TransactionInfo> infoById = PublicMethed
-            .getTransactionInfoById(txid, blockingStubFull);
-    logger.info("Trigger energytotal is " + infoById.get().getReceipt().getEnergyUsageTotal());
-    Assert.assertEquals(
-            Transaction.Result.contractResult.SUCCESS, infoById.get().getReceipt().getResult());
-    return txid;
-  }
-
-  private String triggerTransfer(WalletGrpc.WalletBlockingStub blockingStubFull, byte[] contractAddress,
-                                 byte[] callerAddress, String privateKey, String input) {
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    String txid = PublicMethed.triggerContract(contractAddress,
-            "transfer(bytes32[10][],bytes32[2][],bytes32[9][],bytes32[2],bytes32[21][])",
-            input,
-            true,
-            0L, 1000000000L,
-            callerAddress, privateKey,
-            blockingStubFull);
-    //  logger.info(txid);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Optional<TransactionInfo> infoById = PublicMethed
-            .getTransactionInfoById(txid, blockingStubFull);
-    logger.info("Trigger energytotal is " + infoById.get().getReceipt().getEnergyUsageTotal());
-    Assert.assertEquals(
-            Transaction.Result.contractResult.SUCCESS, infoById.get().getReceipt().getResult());
-    return txid;
-  }
-
-  private String triggerMint(WalletGrpc.WalletBlockingStub blockingStubFull, byte[] contractAddress,
-                             byte[] callerAddress, String privateKey, String input) {
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    String txid = PublicMethed.triggerContract(contractAddress,
-            "mint(uint64,bytes32[9],bytes32[2],bytes32[21])",
-            input,
-            true,
-            0L, 1000000000L,
-            callerAddress, privateKey,
-            blockingStubFull);
-    // logger.info(txid);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Optional<TransactionInfo> infoById = PublicMethed
-            .getTransactionInfoById(txid, blockingStubFull);
-    logger.info("Trigger energytotal is " + infoById.get().getReceipt().getEnergyUsageTotal());
-    Assert.assertEquals(
-            Transaction.Result.contractResult.SUCCESS, infoById.get().getReceipt().getResult());
-    return txid;
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForBurn() throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    long value = 60;
-    String mintInput1 = mintParamsToHexString(mintParam1, value);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setToAmount(60);
-    privateTRC20Builder.setTransparentToAddress(ByteString.copyFrom(callerAddress));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters burnParam = blockingStubFull
-            .createShieldedContractParameters(privateTRC20Builder.build());
-
-    GrpcAPI.PrivateShieldedTRC20Parameters privateTrc20Params = privateTRC20Builder.build();
-    logger.info("input parameters:");
-    logger.info(Hex.toHexString(expsk.getAsk()));
-    logger.info(Hex.toHexString(expsk.getNsk()));
-    for (GrpcAPI.SpendNoteTRC20 spend : privateTrc20Params.getShieldedSpendsList()) {
-      GrpcAPI.Note note = spend.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-      logger.info(Hex.toHexString(spend.getAlpha().toByteArray()));
-      logger.info(Hex.toHexString(spend.getRoot().toByteArray()));
-      logger.info(Hex.toHexString(spend.getPath().toByteArray()));
-      logger.info(String.valueOf(spend.getPos()));
-    }
-    for (GrpcAPI.ReceiveNote rNote : privateTrc20Params.getShieldedReceivesList()) {
-      GrpcAPI.Note note = rNote.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-    }
-    logger.info(String.valueOf(privateTrc20Params.getToAmount()));
-    logger.info(Hex.toHexString(privateTrc20Params.getTransparentToAddress().toByteArray()));
-    logger.info(Hex.toHexString(privateTrc20Params.getShieldedTRC20ContractAddress()
-            .toByteArray()));
-
-    //check the proof
-    boolean result;
-    //verify spendProof && bindingSignature
-    long ctx = JLibrustzcash.librustzcashSaplingVerificationCtxInit();
-    ShieldContract.SpendDescription spend = burnParam.getSpendDescription(0);
-    try {
-      result = JLibrustzcash.librustzcashSaplingCheckSpend(
-              new LibrustzcashParam.CheckSpendParams(ctx,
-                      spend.getValueCommitment().toByteArray(),
-                      spend.getAnchor().toByteArray(),
-                      spend.getNullifier().toByteArray(),
-                      spend.getRk().toByteArray(),
-                      spend.getZkproof().toByteArray(),
-                      spend.getSpendAuthoritySignature().toByteArray(),
-                      burnParam.getMessageHash().toByteArray()));
-      long valueBalance = value;
-      result &= JLibrustzcash.librustzcashSaplingFinalCheck(
-              new LibrustzcashParam.FinalCheckParams(ctx, valueBalance,
-                      burnParam.getBindingSignature().toByteArray(),
-                      burnParam.getMessageHash().toByteArray()));
-    } catch (Throwable any) {
-      result = false;
-    } finally {
-      JLibrustzcash.librustzcashSaplingVerificationCtxFree(ctx);
-    }
-    Assert.assertTrue(result);
-    String burnInput = burnParamsToHexString(burnParam, value, callerAddress);
-    String txid2 = triggerBurn(
-            blockingStubFull, contractAddress, callerAddress, privateKey, burnInput);
-    byte[] nf = burnParam.getSpendDescription(0).getNullifier().toByteArray();
-    logger.info("..............burn result...........");
-    logger.info(txid2);
-    logger.info(Hex.toHexString(nf));
-    logger.info("..............end..............");
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForMintWithoutAsk() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-    long from_amount = 50;
-    SpendingKey sk = SpendingKey.random();
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = expsk.fullViewingKey();
-    byte[] ovk = expsk.getOvk();
-
-    //ReceiveNote
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 50;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    byte[] contractAddress = WalletClient.decodeFromBase58Check(getContractAddress());
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder paramBuilder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    paramBuilder.setOvk(ByteString.copyFrom(ovk));
-    paramBuilder.setFromAmount(from_amount);
-    paramBuilder.addShieldedReceives(revNoteBuilder.build());
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters trc20MintParams = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(paramBuilder.build());
-
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk privateParams = paramBuilder.build();
-    logger.info("input parameters:");
-    logger.info(Hex.toHexString(expsk.getOvk()));
-    logger.info(String.valueOf(privateParams.getFromAmount()));
-    for (GrpcAPI.SpendNoteTRC20 spend : privateParams.getShieldedSpendsList()) {
-      GrpcAPI.Note note = spend.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-      logger.info(Hex.toHexString(spend.getAlpha().toByteArray()));
-      logger.info(Hex.toHexString(spend.getRoot().toByteArray()));
-      logger.info(Hex.toHexString(spend.getPath().toByteArray()));
-      logger.info(String.valueOf(spend.getPos()));
-    }
-    for (GrpcAPI.ReceiveNote rNote : privateParams.getShieldedReceivesList()) {
-      GrpcAPI.Note note = rNote.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-    }
-    logger.info(Hex.toHexString(privateParams.getShieldedTRC20ContractAddress().toByteArray()));
-
-    //verify receiveProof && bindingSignature
-    boolean result;
-    long ctx = JLibrustzcash.librustzcashSaplingVerificationCtxInit();
-    ShieldContract.ReceiveDescription revDesc = trc20MintParams.getReceiveDescription(0);
-    try {
-      result = JLibrustzcash.librustzcashSaplingCheckOutput(
-              new LibrustzcashParam.CheckOutputParams(
-                      ctx,
-                      revDesc.getValueCommitment().toByteArray(),
-                      revDesc.getNoteCommitment().toByteArray(),
-                      revDesc.getEpk().toByteArray(),
-                      revDesc.getZkproof().toByteArray()));
-
-      long valueBalance = -revValue;
-      result &= JLibrustzcash.librustzcashSaplingFinalCheck(
-              new LibrustzcashParam.FinalCheckParams(
-                      ctx,
-                      valueBalance,
-                      trc20MintParams.getBindingSignature().toByteArray(),
-                      trc20MintParams.getMessageHash().toByteArray()));
-    } catch (Throwable any) {
-      result = false;
-    } finally {
-      JLibrustzcash.librustzcashSaplingVerificationCtxFree(ctx);
-    }
-    Assert.assertTrue(result);
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransferWithoutAsk2v2() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam2 = mintParams(
-            privateKey, 40, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam2 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam2);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String mintInput2 = mintParamsToHexString(mintParam2, 40);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    String txid2 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput2);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info(txid2);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    // SpendNoteTRC20 2
-    Optional<TransactionInfo> infoById2 = PublicMethed
-            .getTransactionInfoById(txid2, blockingStubFull);
-    byte[] tx2Data = infoById2.get().getLog(0).getData().toByteArray();
-    long pos2 = bytes32Tolong(ByteArray.subArray(tx2Data, 0, 32));
-    String txid4 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos2);
-    Optional<TransactionInfo> infoById4 = PublicMethed
-            .getTransactionInfoById(txid4, blockingStubFull);
-    byte[] contractResult2 = infoById4.get().getContractResult(0).toByteArray();
-    byte[] path2 = ByteArray.subArray(contractResult2, 32, 1056);
-    byte[] root2 = ByteArray.subArray(contractResult2, 0, 32);
-    logger.info(Hex.toHexString(contractResult2));
-    GrpcAPI.SpendNoteTRC20.Builder note2Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note2Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note2Builder.setPos(pos2);
-    note2Builder.setPath(ByteString.copyFrom(path2));
-    note2Builder.setRoot(ByteString.copyFrom(root2));
-    note2Builder.setNote(mintPrivateParam2.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    privateTRC20Builder.addShieldedSpends(note2Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 30;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    //ReceiveNote 2
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder2 = GrpcAPI.ReceiveNote.newBuilder();
-    FullViewingKey fullViewingKey2 = sk.fullViewingKey();
-    IncomingViewingKey incomingViewingKey2 = fullViewingKey2.inViewingKey();
-    PaymentAddress paymentAddress2 = incomingViewingKey2.address(DiversifierT.random()).get();
-    long revValue2 = 70;
-    byte[] memo2 = new byte[512];
-    byte[] rcm2 = Note.generateR();
-    String paymentAddressStr2 = KeyIo.encodePaymentAddress(paymentAddress2);
-    GrpcAPI.Note revNote2 = getNote(revValue2, paymentAddressStr2, rcm2, memo2);
-    revNoteBuilder2.setNote(revNote2);
-
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder2.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    privateTRC20Builder.setAk(ByteString.copyFrom(fvk.getAk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-
-    //logger.info(privateTRC20Builder.build().toString());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(privateTRC20Builder.build());
-    logger.info(transferParam.toString());
-    // checkTransferParams(transferParam);
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk privateParams = privateTRC20Builder.build();
-    logger.info("input parameters:");
-    logger.info(Hex.toHexString(expsk.getNsk()));
-    logger.info(Hex.toHexString(fvk.getAk()));
-    logger.info(Hex.toHexString(expsk.getOvk()));
-    //logger.info(String.valueOf(privateParams.getFromAmount()));
-    for (GrpcAPI.SpendNoteTRC20 spend : privateParams.getShieldedSpendsList()) {
-      GrpcAPI.Note note = spend.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-      logger.info(Hex.toHexString(spend.getAlpha().toByteArray()));
-      logger.info(Hex.toHexString(spend.getRoot().toByteArray()));
-      logger.info(Hex.toHexString(spend.getPath().toByteArray()));
-      logger.info(String.valueOf(spend.getPos()));
-    }
-    for (GrpcAPI.ReceiveNote rNote : privateParams.getShieldedReceivesList()) {
-      GrpcAPI.Note note = rNote.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-    }
-    logger.info(Hex.toHexString(privateParams.getShieldedTRC20ContractAddress().toByteArray()));
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransferWithoutAsk2v1() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam2 = mintParams(
-            privateKey, 40, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam2 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam2);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String mintInput2 = mintParamsToHexString(mintParam2, 40);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    String txid2 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput2);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info(txid2);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    // SpendNoteTRC20 2
-    Optional<TransactionInfo> infoById2 = PublicMethed
-            .getTransactionInfoById(txid2, blockingStubFull);
-    byte[] tx2Data = infoById2.get().getLog(0).getData().toByteArray();
-    long pos2 = bytes32Tolong(ByteArray.subArray(tx2Data, 0, 32));
-    String txid4 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos2);
-    Optional<TransactionInfo> infoById4 = PublicMethed
-            .getTransactionInfoById(txid4, blockingStubFull);
-    byte[] contractResult2 = infoById4.get().getContractResult(0).toByteArray();
-    byte[] path2 = ByteArray.subArray(contractResult2, 32, 1056);
-    byte[] root2 = ByteArray.subArray(contractResult2, 0, 32);
-    logger.info(Hex.toHexString(contractResult2));
-    GrpcAPI.SpendNoteTRC20.Builder note2Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note2Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note2Builder.setPos(pos2);
-    note2Builder.setPath(ByteString.copyFrom(path2));
-    note2Builder.setRoot(ByteString.copyFrom(root2));
-    note2Builder.setNote(mintPrivateParam2.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    privateTRC20Builder.addShieldedSpends(note2Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 100;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    privateTRC20Builder.setAk(ByteString.copyFrom(fvk.getAk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(privateTRC20Builder.build());
-    logger.info(transferParam.toString());
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransferWithoutAsk1v1() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull
-            .createShieldedContractParameters(mintPrivateParam1);
-    String mintInput1 = mintParamsToHexString(mintParam1, 60);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(
-            blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 100;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    privateTRC20Builder.setAk(ByteString.copyFrom(fvk.getAk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(privateTRC20Builder.build());
-    logger.info(transferParam.toString());
-  }
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForTransferWithoutAsk1v2() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 100, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = blockingStubFull.createShieldedContractParameters(
-            mintPrivateParam1);
-
-    String mintInput1 = mintParamsToHexString(mintParam1, 100);
-    String txid1 = triggerMint(blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-
-    //ReceiveNote 1
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder = GrpcAPI.ReceiveNote.newBuilder();
-    SpendingKey spendingKey = SpendingKey.random();
-    FullViewingKey fullViewingKey = spendingKey.fullViewingKey();
-    IncomingViewingKey incomingViewingKey = fullViewingKey.inViewingKey();
-    PaymentAddress paymentAddress = incomingViewingKey.address(DiversifierT.random()).get();
-    long revValue = 30;
-    byte[] memo = new byte[512];
-    byte[] rcm = Note.generateR();
-    String paymentAddressStr = KeyIo.encodePaymentAddress(paymentAddress);
-    GrpcAPI.Note revNote = getNote(revValue, paymentAddressStr, rcm, memo);
-    revNoteBuilder.setNote(revNote);
-
-    //ReceiveNote 2
-    GrpcAPI.ReceiveNote.Builder revNoteBuilder2 = GrpcAPI.ReceiveNote.newBuilder();
-    FullViewingKey fullViewingKey2 = sk.fullViewingKey();
-    IncomingViewingKey incomingViewingKey2 = fullViewingKey2.inViewingKey();
-    PaymentAddress paymentAddress2 = incomingViewingKey2.address(DiversifierT.random()).get();
-    long revValue2 = 70;
-    byte[] memo2 = new byte[512];
-    byte[] rcm2 = Note.generateR();
-    String paymentAddressStr2 = KeyIo.encodePaymentAddress(paymentAddress2);
-    GrpcAPI.Note revNote2 = getNote(revValue2, paymentAddressStr2, rcm2, memo2);
-    revNoteBuilder2.setNote(revNote2);
-
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder.build());
-    privateTRC20Builder.addShieldedReceives(revNoteBuilder2.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    privateTRC20Builder.setAk(ByteString.copyFrom(fvk.getAk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(privateTRC20Builder.build());
-    logger.info(transferParam.toString());
-  }
-
-
-  @Ignore
-  @Test
-  public void testCreateShieldedContractParametersForBurnWithoutAsk() throws Exception {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, 60, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-            mintPrivateParam1);
-    long value = 60;
-    String mintInput1 = mintParamsToHexString(mintParam1, value);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20ParametersWithoutAsk.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    FullViewingKey fvk = expsk.fullViewingKey();
-    privateTRC20Builder.setAk(ByteString.copyFrom(fvk.getAk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    //privateTRC20Builder.setOvk(ByteString.copyFrom(expsk.getOvk()));
-    privateTRC20Builder.setToAmount(60);
-    privateTRC20Builder.setTransparentToAddress(ByteString.copyFrom(callerAddress));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    //logger.info(privateTRC20Builder.build().toString());
-    GrpcAPI.ShieldedTRC20Parameters transferParam = blockingStubFull
-            .createShieldedContractParametersWithoutAsk(privateTRC20Builder.build());
-    logger.info(transferParam.toString());
-
-    GrpcAPI.PrivateShieldedTRC20ParametersWithoutAsk privateParams = privateTRC20Builder.build();
-    logger.info("input parameters:");
-    logger.info(Hex.toHexString(expsk.getNsk()));
-    logger.info(Hex.toHexString(fvk.getAk()));
-    //logger.info(Hex.toHexString(expsk.getOvk()));
-    //logger.info(String.valueOf(privateParams.getFromAmount()));
-    for (GrpcAPI.SpendNoteTRC20 spend : privateParams.getShieldedSpendsList()) {
-      GrpcAPI.Note note = spend.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-      logger.info(Hex.toHexString(spend.getAlpha().toByteArray()));
-      logger.info(Hex.toHexString(spend.getRoot().toByteArray()));
-      logger.info(Hex.toHexString(spend.getPath().toByteArray()));
-      logger.info(String.valueOf(spend.getPos()));
-    }
-    for (GrpcAPI.ReceiveNote rNote : privateParams.getShieldedReceivesList()) {
-      GrpcAPI.Note note = rNote.getNote();
-      logger.info(String.valueOf(note.getValue()));
-      logger.info(note.getPaymentAddress());
-      logger.info(Hex.toHexString(note.getRcm().toByteArray()));
-    }
-    logger.info(String.valueOf(privateParams.getToAmount()));
-    logger.info(Hex.toHexString(privateParams.getTransparentToAddress().toByteArray()));
-    logger.info(Hex.toHexString(privateParams.getShieldedTRC20ContractAddress().toByteArray()));
-  }
-
-  @Ignore
-  @Test
-  public void testScanShieldedTRC20NotesbyIvk() throws ZksnarkException {
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    int statNum = 1;
-    int endNum = 140;
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(getContractAddress());
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    ExpandedSpendingKey esk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    byte[] ivk = fvk.inViewingKey().value;
-    GrpcAPI.IvkDecryptTRC20Parameters.Builder paramBuilder = GrpcAPI
-            .IvkDecryptTRC20Parameters.newBuilder();
-    logger.info(Hex.toHexString(contractAddress));
-    logger.info(Hex.toHexString(ivk));
-    logger.info(Hex.toHexString(fvk.getAk()));
-    logger.info(Hex.toHexString(fvk.getNk()));
-
-    paramBuilder.setAk(ByteString.copyFrom(fvk.getAk()));
-    paramBuilder.setNk(ByteString.copyFrom(fvk.getNk()));
-    paramBuilder.setIvk(ByteString.copyFrom(ivk));
-    paramBuilder.setStartBlockIndex(statNum);
-    paramBuilder.setEndBlockIndex(endNum);
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.DecryptNotesTRC20 scannedNotes = blockingStubFull.scanShieldedTRC20NotesbyIvk(
-            paramBuilder.build());
-    for (GrpcAPI.DecryptNotesTRC20.NoteTx noteTx : scannedNotes.getNoteTxsList()) {
-      logger.info(noteTx.toString());
-    }
-  }
-
-  @Ignore
-  @Test
-  public void testscanShieldedTRC20NotesbyOvk() throws ZksnarkException {
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    int statNum = 1;
-    int endNum = 140;
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(getContractAddress());
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    FullViewingKey fvk = sk.fullViewingKey();
-    GrpcAPI.OvkDecryptTRC20Parameters.Builder paramBuilder = GrpcAPI.OvkDecryptTRC20Parameters
-            .newBuilder();
-    logger.info(Hex.toHexString(contractAddress));
-    logger.info(Hex.toHexString(fvk.getOvk()));
-
-    paramBuilder.setOvk(ByteString.copyFrom(fvk.getOvk()));
-    paramBuilder.setStartBlockIndex(statNum);
-    paramBuilder.setEndBlockIndex(endNum);
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.DecryptNotesTRC20 scannedNotes = blockingStubFull.scanShieldedTRC20NotesbyOvk(
-            paramBuilder.build());
-    for (GrpcAPI.DecryptNotesTRC20.NoteTx noteTx : scannedNotes.getNoteTxsList()) {
-      logger.info(noteTx.toString());
-    }
-  }
-
-  @Ignore
-  @Test
-  public void isShieldedTRC20ContractNoteSpent() throws ZksnarkException {
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    int statNum = 770;
-    int endNum = 870;
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(getContractAddress());
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    ExpandedSpendingKey esk = sk.expandedSpendingKey();
-    FullViewingKey fvk = sk.fullViewingKey();
-    byte[] ivk = fvk.inViewingKey().value;
-    GrpcAPI.IvkDecryptTRC20Parameters.Builder paramBuilder = GrpcAPI.IvkDecryptTRC20Parameters
-            .newBuilder();
-
-    paramBuilder.setAk(ByteString.copyFrom(fvk.getAk()));
-    paramBuilder.setNk(ByteString.copyFrom(fvk.getNk()));
-    paramBuilder.setIvk(ByteString.copyFrom(ivk));
-    paramBuilder.setStartBlockIndex(statNum);
-    paramBuilder.setEndBlockIndex(endNum);
-    paramBuilder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.DecryptNotesTRC20 scannedNotes = blockingStubFull.scanShieldedTRC20NotesbyIvk(
-            paramBuilder.build());
-    for (GrpcAPI.DecryptNotesTRC20.NoteTx noteTx : scannedNotes.getNoteTxsList()) {
-      logger.info(noteTx.toString());
-    }
-
-    GrpcAPI.NfTRC20Parameters.Builder NfBuilfer;
-    NfBuilfer = GrpcAPI.NfTRC20Parameters.newBuilder();
-    NfBuilfer.setAk(ByteString.copyFrom(fvk.getAk()));
-    NfBuilfer.setNk(ByteString.copyFrom(fvk.getNk()));
-    NfBuilfer.setPosition(0);
-    NfBuilfer.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    NfBuilfer.setNote(scannedNotes.getNoteTxs(0).getNote());
-    GrpcAPI.NullifierResult result = blockingStubFull.isShieldedTRC20ContractNoteSpent(
-            NfBuilfer.build());
-    Assert.assertTrue(result.getIsSpent());
-
-    NfBuilfer = GrpcAPI.NfTRC20Parameters.newBuilder();
-    NfBuilfer.setAk(ByteString.copyFrom(fvk.getAk()));
-    NfBuilfer.setNk(ByteString.copyFrom(fvk.getNk()));
-    NfBuilfer.setPosition(1);
-    NfBuilfer.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    NfBuilfer.setNote(scannedNotes.getNoteTxs(1).getNote());
-    GrpcAPI.NullifierResult result1 = blockingStubFull.isShieldedTRC20ContractNoteSpent(
-            NfBuilfer.build());
-    Assert.assertTrue(result1.getIsSpent());
-    GrpcAPI.NfTRC20Parameters nfParma = NfBuilfer.build();
-    logger.info(String.valueOf(nfParma.getNote().getValue()));
-    logger.info(nfParma.getNote().getPaymentAddress());
-    logger.info(Hex.toHexString(nfParma.getNote().getRcm().toByteArray()));
-    logger.info(Hex.toHexString(nfParma.getNote().getMemo().toByteArray()));
-    logger.info(Hex.toHexString(nfParma.getAk().toByteArray()));
-    logger.info(Hex.toHexString(nfParma.getNk().toByteArray()));
-    logger.info(String.valueOf(nfParma.getPosition()));
-    logger.info(Hex.toHexString(nfParma.getShieldedTRC20ContractAddress().toByteArray()));
-
-    NfBuilfer = GrpcAPI.NfTRC20Parameters.newBuilder();
-    NfBuilfer.setAk(ByteString.copyFrom(fvk.getAk()));
-    NfBuilfer.setNk(ByteString.copyFrom(fvk.getNk()));
-    NfBuilfer.setPosition(3);
-    NfBuilfer.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    NfBuilfer.setNote(scannedNotes.getNoteTxs(2).getNote());
-    GrpcAPI.NullifierResult result2 = blockingStubFull.isShieldedTRC20ContractNoteSpent(
-            NfBuilfer.build());
-    Assert.assertFalse(result2.getIsSpent());
-  }
-
-  @Ignore
-  @Test
-  public void testTriggerNullifer() throws ZksnarkException, ContractValidateException {
-    librustzcashInitZksnarkParams();
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    SpendingKey sk = SpendingKey.decode(privateKey);
-    //String  privateKey2 = "03d51abbd89cb8196f0efb6892f94d68fccc2c35f0b84609e5f12c55dd85aba8";
-    long value = 60;
-    GrpcAPI.PrivateShieldedTRC20Parameters mintPrivateParam1 = mintParams(
-            privateKey, value, contractAddr);
-    GrpcAPI.ShieldedTRC20Parameters mintParam1 = wallet.createShieldedContractParameters(
-            mintPrivateParam1);
-    String mintInput1 = mintParamsToHexString(mintParam1, value);
-    String txid1 = triggerMint(
-            blockingStubFull, contractAddress, callerAddress, privateKey, mintInput1);
-    logger.info("..............min result...........");
-    logger.info(txid1);
-    logger.info("..............end..............");
-
-    // SpendNoteTRC20 1
-    Optional<TransactionInfo> infoById1 = PublicMethed
-            .getTransactionInfoById(txid1, blockingStubFull);
-    byte[] tx1Data = infoById1.get().getLog(0).getData().toByteArray();
-    long pos1 = bytes32Tolong(ByteArray.subArray(tx1Data, 0, 32));
-    String txid3 = triggerGetPath(blockingStubFull, contractAddress, callerAddress, privateKey, pos1);
-    Optional<TransactionInfo> infoById3 = PublicMethed
-            .getTransactionInfoById(txid3, blockingStubFull);
-    byte[] contractResult1 = infoById3.get().getContractResult(0).toByteArray();
-    byte[] path1 = ByteArray.subArray(contractResult1, 32, 1056);
-    byte[] root1 = ByteArray.subArray(contractResult1, 0, 32);
-    logger.info(Hex.toHexString(contractResult1));
-    GrpcAPI.SpendNoteTRC20.Builder note1Builder = GrpcAPI.SpendNoteTRC20.newBuilder();
-    note1Builder.setAlpha(ByteString.copyFrom(Note.generateR()));
-    note1Builder.setPos(pos1);
-    note1Builder.setPath(ByteString.copyFrom(path1));
-    note1Builder.setRoot(ByteString.copyFrom(root1));
-    note1Builder.setNote(mintPrivateParam1.getShieldedReceives(0).getNote());
-
-    GrpcAPI.PrivateShieldedTRC20Parameters.Builder privateTRC20Builder = GrpcAPI
-            .PrivateShieldedTRC20Parameters.newBuilder();
-    privateTRC20Builder.addShieldedSpends(note1Builder.build());
-    ExpandedSpendingKey expsk = sk.expandedSpendingKey();
-    privateTRC20Builder.setAsk(ByteString.copyFrom(expsk.getAsk()));
-    privateTRC20Builder.setNsk(ByteString.copyFrom(expsk.getNsk()));
-    privateTRC20Builder.setToAmount(60);
-    privateTRC20Builder.setTransparentToAddress(ByteString.copyFrom(callerAddress));
-    privateTRC20Builder.setShieldedTRC20ContractAddress(ByteString.copyFrom(contractAddress));
-    GrpcAPI.ShieldedTRC20Parameters burnParam = wallet.createShieldedContractParameters(privateTRC20Builder.build());
-    String burnInput = burnParamsToHexString(burnParam, value, callerAddress);
-    String txid2 = triggerBurn(blockingStubFull, contractAddress, callerAddress, privateKey, burnInput);
-    byte[] nf = burnParam.getSpendDescription(0).getNullifier().toByteArray();
-    logger.info("..............burn result...........");
-    logger.info(txid2);
-    logger.info(Hex.toHexString(nf));
-    logger.info("..............end..............");
-
-    //test nullifer
-    String methodSign = "nullifiers(bytes32)";
-    byte[] selector = new byte[4];
-    System.arraycopy(Hash.sha3(methodSign.getBytes()), 0, selector, 0, 4);
-    byte[] input = ByteUtil.merge(selector, nf);
-    SmartContractOuterClass.TriggerSmartContract.Builder triggerBuilder = SmartContractOuterClass
-            .TriggerSmartContract.newBuilder();
-    triggerBuilder.setContractAddress(ByteString.copyFrom(contractAddress));
-    triggerBuilder.setData(ByteString.copyFrom(input));
-    GrpcAPI.TransactionExtention trxExt2 = blockingStubFull.triggerConstantContract(
-            triggerBuilder.build());
-    String code = trxExt2.getResult().getCode().toString();
-    boolean bool = trxExt2.getResult().getResult();
-    List<ByteString> list = trxExt2.getConstantResultList();
-    //String message = list.toArray().toString();
-    byte[] listBytes = new byte[0];
-    for (ByteString bs : list) {
-      listBytes = ByteUtil.merge(listBytes, bs.toByteArray());
-    }
-    logger.info("..............nullifier result...........");
-    logger.info(code);
-    logger.info(String.valueOf(bool));
-    logger.info(Hex.toHexString(listBytes));
-    logger.info("..............end..............");
-    Assert.assertArrayEquals(nf, listBytes);
-  }
-
-  @Ignore
-  @Test
-  public void testNullier() {
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    byte[] nf = Hex.decode("5f8c5fe4948b86bf7acf230925afb9b073b03f97231d95cb09cb01ae3eec5a60");
-    String methodSign = "nullifiers(bytes32)";
-    byte[] selector = new byte[4];
-    System.arraycopy(Hash.sha3(methodSign.getBytes()), 0, selector, 0, 4);
-    byte[] input = ByteUtil.merge(selector, nf);
-    SmartContractOuterClass.TriggerSmartContract.Builder triggerBuilder = SmartContractOuterClass
-            .TriggerSmartContract.newBuilder();
-    String contractAddr = getContractAddress();
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check(contractAddr);
-    triggerBuilder.setContractAddress(ByteString.copyFrom(contractAddress));
-    triggerBuilder.setData(ByteString.copyFrom(input));
-    GrpcAPI.TransactionExtention trxExt2 = blockingStubFull.triggerConstantContract(
-            triggerBuilder.build());
-
-    String code = trxExt2.getResult().getCode().toString();
-    boolean bool = trxExt2.getResult().getResult();
-    List<ByteString> list = trxExt2.getConstantResultList();
-    //String message = list.toArray().toString();
-    byte[] listBytes = new byte[0];
-    for (ByteString bs : list) {
-      listBytes = ByteUtil.merge(listBytes, bs.toByteArray());
-    }
-    logger.info("..............nullifier result...........");
-    logger.info(code);
-    logger.info(String.valueOf(bool));
-    logger.info(Hex.toHexString(listBytes));
-    logger.info("..............end..............");
-  }
-
-  private String getContractAddress() {
-    return "TX29caJFwDPZ9tzjuQ1GB6Ci59ocxQThKN";
-  }
-
-  private String mintParamsToHexString(GrpcAPI.ShieldedTRC20Parameters mintParams, long value) {
-    byte[] mergedBytes;
-    ShieldContract.ReceiveDescription revDesc = mintParams.getReceiveDescription(0);
-    mergedBytes = ByteUtil.merge(
-            longTo32Bytes(value),
-            revDesc.getNoteCommitment().toByteArray(),
-            revDesc.getValueCommitment().toByteArray(),
-            revDesc.getEpk().toByteArray(),
-            revDesc.getZkproof().toByteArray(),
-            mintParams.getBindingSignature().toByteArray(),
-            revDesc.getCEnc().toByteArray(),
-            revDesc.getCOut().toByteArray(),
-            new byte[12]
-    );
-    return Hex.toHexString(mergedBytes);
-  }
-
-  private String transferParamsToHexString(GrpcAPI.ShieldedTRC20Parameters params) {
-    byte[] input = new byte[0];
-    byte[] spendAuthSig = new byte[0];
-    byte[] output = new byte[0];
-    byte[] c = new byte[0];
-    byte[] bindingSig;
-    byte[] mergedBytes;
-    List<ShieldContract.SpendDescription> spendDescs = params.getSpendDescriptionList();
-    for (ShieldContract.SpendDescription spendDesc : spendDescs) {
-      input = ByteUtil.merge(input,
-              spendDesc.getNullifier().toByteArray(),
-              spendDesc.getAnchor().toByteArray(),
-              spendDesc.getValueCommitment().toByteArray(),
-              spendDesc.getRk().toByteArray(),
-              spendDesc.getZkproof().toByteArray()
-      );
-      spendAuthSig = ByteUtil.merge(
-              spendAuthSig, spendDesc.getSpendAuthoritySignature().toByteArray());
-    }
-    byte[] inputOffsetbytes = longTo32Bytes(192);
-    long spendCount = spendDescs.size();
-    byte[] spendCountBytes = longTo32Bytes(spendCount);
-    byte[] authOffsetBytes = longTo32Bytes(192 + 32 + 320 * spendCount);
-    List<ShieldContract.ReceiveDescription> recvDescs = params.getReceiveDescriptionList();
-    for (ShieldContract.ReceiveDescription recvDesc : recvDescs) {
-      output = ByteUtil.merge(output,
-              recvDesc.getNoteCommitment().toByteArray(),
-              recvDesc.getValueCommitment().toByteArray(),
-              recvDesc.getEpk().toByteArray(),
-              recvDesc.getZkproof().toByteArray()
-      );
-      c = ByteUtil.merge(c,
-              recvDesc.getCEnc().toByteArray(),
-              recvDesc.getCOut().toByteArray(),
-              new byte[12]
-      );
-    }
-    long recvCount = recvDescs.size();
-    byte[] recvCountBytes = longTo32Bytes(recvCount);
-    byte[] outputOffsetbytes = longTo32Bytes(192 + 32 + 320 * spendCount + 32 + 64 * spendCount);
-    byte[] coffsetBytes = longTo32Bytes(192 + 32 + 320 * spendCount + 32 + 64 * spendCount + 32
-            + 288 * recvCount);
-    bindingSig = params.getBindingSignature().toByteArray();
-    mergedBytes = ByteUtil.merge(inputOffsetbytes,
-            authOffsetBytes,
-            outputOffsetbytes,
-            bindingSig,
-            coffsetBytes,
-            spendCountBytes,
-            input,
-            spendCountBytes,
-            spendAuthSig,
-            recvCountBytes,
-            output,
-            recvCountBytes,
-            c
-    );
-    //logger.info(ByteArray.toHexString(mergedBytes));
-    return Hex.toHexString(mergedBytes);
-  }
-
-  private String burnParamsToHexString(GrpcAPI.ShieldedTRC20Parameters burnParams, long value,
-                                       byte[] transparentToAddress) {
-    byte[] mergedBytes;
-    byte[] payTo = new byte[32];
-    System.arraycopy(transparentToAddress, 0, payTo, 11, 21);
-    ShieldContract.SpendDescription spendDesc = burnParams.getSpendDescription(0);
-    mergedBytes = ByteUtil.merge(
-            spendDesc.getNullifier().toByteArray(),
-            spendDesc.getAnchor().toByteArray(),
-            spendDesc.getValueCommitment().toByteArray(),
-            spendDesc.getRk().toByteArray(),
-            spendDesc.getZkproof().toByteArray(),
-            spendDesc.getSpendAuthoritySignature().toByteArray(),
-            longTo32Bytes(value),
-            burnParams.getBindingSignature().toByteArray(),
-            payTo
-    );
-    logger.info("merged bytes: " + ByteArray.toHexString(mergedBytes));
-    return Hex.toHexString(mergedBytes);
-  }
-
-  private byte[] longTo32Bytes(long value) {
-    byte[] longBytes = ByteArray.fromLong(value);
-    byte[] zeroBytes = new byte[24];
-    return ByteUtil.merge(zeroBytes, longBytes);
-  }
-
-  private long bytes32Tolong(byte[] value) {
-    return ByteArray.toLong(value);
-  }
-
-  @Ignore
-  @Test
-  public void getTrc20AccountBalance() {
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check("TG3Tj3hwXMgE33zWugDTh2NWshRexU4QtN");
-    logger.info("trc20 contract address: " + ByteArray.toHexString(contractAddress));
-    ManagedChannel channelFull = null;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    byte[] userAccountAddress = new byte[32];
-    String shieldedContractAddr = getContractAddress();
-    byte[] shieldedContractAddress = WalletClient
-            .decodeFromBase58Check(shieldedContractAddr);
-    System.arraycopy(shieldedContractAddress, 0, userAccountAddress, 11, 21);
-
-    String methodSign = "balanceOf(address)";
-    byte[] selector = new byte[4];
-    System.arraycopy(Hash.sha3(methodSign.getBytes()), 0, selector, 0, 4);
-    byte[] input = ByteUtil.merge(selector, userAccountAddress);
-    SmartContractOuterClass.TriggerSmartContract.Builder triggerBuilder = SmartContractOuterClass
-            .TriggerSmartContract.newBuilder();
-    triggerBuilder.setContractAddress(ByteString.copyFrom(contractAddress));
-    triggerBuilder.setData(ByteString.copyFrom(input));
-    GrpcAPI.TransactionExtention trxExt2 = blockingStubFull.triggerConstantContract(
-            triggerBuilder.build());
-
-    List<ByteString> list = trxExt2.getConstantResultList();
-    byte[] listBytes = new byte[0];
-    for (ByteString bs : list) {
-      listBytes = ByteUtil.merge(listBytes, bs.toByteArray());
-    }
-    logger.info("balance " + Hex.toHexString(listBytes));
-  }
-
-  @Ignore
-  @Test
-  public void setAllowance() {
-    byte[] contractAddress = WalletClient
-            .decodeFromBase58Check("TG3Tj3hwXMgE33zWugDTh2NWshRexU4QtN");
-    ManagedChannel channelFull;
-    WalletGrpc.WalletBlockingStub blockingStubFull = null;
-    String fullnode = "127.0.0.1:50051";
-    channelFull = ManagedChannelBuilder.forTarget(fullnode)
-            .usePlaintext(true)
-            .build();
-    blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
-
-    String shieldedContractAddr = getContractAddress();
-    byte[] shieldedContractAddress = WalletClient
-            .decodeFromBase58Check(shieldedContractAddr);
-    byte[] shieldedContractAddressPadding = new byte[32];
-    System.arraycopy(shieldedContractAddress, 0, shieldedContractAddressPadding, 11, 21);
-    logger.info("shielded contract addr " + ByteArray.toHexString(shieldedContractAddressPadding));
-    byte[] valueBytes = longTo32Bytes(100_000L);
-    String input = Hex.toHexString(ByteUtil.merge(shieldedContractAddressPadding, valueBytes));
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    byte[] callerAddress = WalletClient.decodeFromBase58Check("TFsrP7YcSSRwHzLPwaCnXyTKagHs8rXKNJ");
-    String privateKey = "650950B193DDDDB35B6E48912DD28F7AB0E7140C1BFDEFD493348F02295BD812";
-    String txid = PublicMethed.triggerContract(contractAddress,
-            "approve(address,uint256)",
-            input,
-            true,
-            0L,
-            1000000000L,
-            callerAddress,
-            privateKey,
-            blockingStubFull);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Optional<TransactionInfo> infoById = PublicMethed
-            .getTransactionInfoById(txid, blockingStubFull);
-    Assert.assertEquals(
-            Transaction.Result.contractResult.SUCCESS, infoById.get().getReceipt().getResult());
   }
 }
